+++ PSFramework/functions/utility/ConvertTo-PSFHashtable.ps1
@@ -12,12 +12,21 @@
 		The propertynames to exclude.
 		Must be full property-names, no wildcard/regex matching.
 	
+	.PARAMETER Include
+		The propertynames to include.
+		Must be full property-names, no wildcard/regex matching.
+	
+	.PARAMETER IncludeEmpty
+		By default, only properties on the input object are included.
+		In order to force all properties defiend in -Include to be included, specify this switch.
+		Keys added through this have an empty ($null) value.
+	
 	.PARAMETER InputObject
 		The object(s) to convert
 	
 	.EXAMPLE
 		PS C:\> Get-ChildItem | ConvertTo-PSFHashtable
-	
+		
 		Scans all items in the current path and converts those objects into hashtables.
 #>
 	[OutputType([System.Collections.Hashtable])]
@@ -28,6 +37,9 @@
 		
 		[string[]]
 		$Include,
+		
+		[switch]
+		$IncludeEmpty,
 		
 		[Parameter(ValueFromPipeline = $true)]
 		$InputObject
@@ -48,6 +60,11 @@
 					{
 						if ($key -notin $Include) { $hashTable.Remove($key) }
 					}
+					if (-not $IncludeEmpty) { continue }
+					foreach ($key in $Include)
+					{
+						if ($hashTable.Keys -notcontains $key) { $hashTable[$key] = $null }
+					}
 				}
 				$hashTable
 			}
@@ -60,6 +77,13 @@
 					if ($Include -and ($name -notin $Include)) { continue }
 					$hashTable[$name] = $item[$name]
 				}
+				if ($Include -and $IncludeEmpty)
+				{
+					foreach ($key in $Include)
+					{
+						if ($hashTable.Keys -notcontains $key) { $hashTable[$key] = $null }
+					}
+				}
 				$hashTable
 			}
 			else
@@ -71,6 +95,13 @@
 					if ($Include -and ($property.Name -notin $Include)) { continue }
 					
 					$hashTable[$property.Name] = $property.Value
+				}
+				if ($Include -and $IncludeEmpty)
+				{
+					foreach ($key in $Include)
+					{
+						if ($hashTable.Keys -notcontains $key) { $hashTable[$key] = $null }
+					}
 				}
 				$hashTable
 			}